requirements.txt
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
pymysql
sqlalchemy